Extract copytojson into hook
