+++ x-pack/plugins/observability/public/pages/slo_edit/components/slo_edit_form.tsx
@@ -35,6 +35,7 @@ import {
   CREATE_RULE_SEARCH_PARAM,
   useAddRuleFlyoutState,
 } from '../hooks/use_add_rule_flyout_state';
+import { useCopyToJson } from '../hooks/use_copy_to_json';
 import { useParseUrlState } from '../hooks/use_parse_url_state';
 import { useSectionFormValidation } from '../hooks/use_section_form_validation';
 import { useShowSections } from '../hooks/use_show_sections';
@@ -53,7 +54,6 @@ export function SloEditForm({ slo }: Props) {
   const {
     application: { navigateToUrl },
     http: { basePath },
-    notifications,
     triggersActionsUi: { getAddRuleFlyout: AddRuleFlyout },
   } = useKibana().services;
 
@@ -78,6 +78,7 @@ export function SloEditForm({ slo }: Props) {
     mode: 'all',
   });
   const { watch, getFieldState, getValues, formState, trigger } = methods;
+  const handleCopyToJson = useCopyToJson({ trigger, getValues });
 
   const { isIndicatorSectionValid, isObjectiveSectionValid, isDescriptionSectionValid } =
     useSectionFormValidation({
@@ -96,35 +97,6 @@ export function SloEditForm({ slo }: Props) {
 
   const { mutateAsync: createSlo, isLoading: isCreateSloLoading } = useCreateSlo();
   const { mutateAsync: updateSlo, isLoading: isUpdateSloLoading } = useUpdateSlo();
-
-  const handleCopyToJson = async () => {
-    const isValid = await trigger();
-    if (!isValid) {
-      return;
-    }
-    const values = transformCreateSLOFormToCreateSLOInput(getValues());
-    try {
-      await copyTextToClipboard(JSON.stringify(values, null, 2));
-      notifications.toasts.add({
-        title: i18n.translate('xpack.observability.slo.sloEdit.copyJsonNotification', {
-          defaultMessage: 'JSON copied to clipboard',
-        }),
-      });
-    } catch (e) {
-      notifications.toasts.add({
-        title: i18n.translate('xpack.observability.slo.sloEdit.copyJsonFailedNotification', {
-          defaultMessage: 'Could not copy JSON to clipboard',
-        }),
-      });
-    }
-  };
-
-  const copyTextToClipboard = async (text: string) => {
-    if (!window.navigator?.clipboard) {
-      throw new Error('Could not copy to clipboard!');
-    }
-    await window.navigator.clipboard.writeText(text);
-  };
 
   const handleSubmit = async () => {
     const isValid = await trigger();
